+++ test/test_eigen.py
@@ -15,10 +15,12 @@ for p in inc_paths:
 @mark.skipif(eigen_path is None, reason="Eigen not found")
 class TestEIGEN:
     def setup_class(cls):
-        import cppyy
+        import cppyy, warnings
 
         cppyy.add_include_path(eigen_path)
-        cppyy.include('Eigen/Dense')
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore')
+            cppyy.include('Eigen/Dense')
 
     def test01_simple_matrix_and_vector(self):
         """Basic creation of an Eigen::Matrix and Eigen::Vector"""
@@ -145,10 +147,12 @@ class TestEIGEN:
 @mark.skipif(eigen_path is None, reason="Eigen not found")
 class TestEIGEN_REGRESSIOn:
     def setup_class(cls):
-        import cppyy
+        import cppyy, warnings
 
         cppyy.add_include_path(eigen_path)
-        cppyy.include('Eigen/Dense')
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore')
+            cppyy.include('Eigen/Dense')
 
     def test01_use_of_Map(self):
         """Use of Map (used to crash)"""