+++ doc/source/changelog.rst
@@ -16,6 +16,7 @@ master: 3.0.0
 * Upgrade backend to Cling on top of LLVM 13
 * Improve handling of `const char*` as template argument
 * Fix regression in use of unnamed but typedef'ed enums
+* Report C++ warnings from ``cppdef`` as ``SyntaxWarning``
 
 
 2023-01-21: 2.4.2